# Reuse iterator in `batch.NewChunkMergeIterator`