--- a/pkg/querier/batch/chunk_test.go
+++ b/pkg/querier/batch/chunk_test.go
@@ -35,10 +35,10 @@ func testChunkIter(t *testing.T, encoding chunk.Encoding) {
 	iter := &chunkIterator{}
 
 	iter.reset(chunk)
-	testIter(t, 100, newIteratorAdapter(iter), encoding)
+	testIter(t, 100, newIteratorAdapter(nil, iter), encoding)
 
 	iter.reset(chunk)
-	testSeek(t, 100, newIteratorAdapter(iter), encoding)
+	testSeek(t, 100, newIteratorAdapter(nil, iter), encoding)
 }
 
 func mkChunk(t require.TestingT, from model.Time, points int, encoding chunk.Encoding) chunk.Chunk {