--- a/pkg/querier/distributor_queryable_streaming.go
+++ b/pkg/querier/distributor_queryable_streaming.go
@@ -29,7 +29,7 @@ func (s *streamingChunkSeries) Labels() labels.Labels {
 	return s.labels
 }
 
-func (s *streamingChunkSeries) Iterator(_ chunkenc.Iterator) chunkenc.Iterator {
+func (s *streamingChunkSeries) Iterator(it chunkenc.Iterator) chunkenc.Iterator {
 	var uniqueChunks []client.Chunk
 	totalChunks := 0
 
@@ -62,5 +62,5 @@ func (s *streamingChunkSeries) Iterator(_ chunkenc.Iterator) chunkenc.Iterator {
 		return series.NewErrIterator(err)
 	}
 
-	return s.chunkIteratorFunc(chunks, model.Time(s.mint), model.Time(s.maxt))
+	return s.chunkIteratorFunc(it, chunks, model.Time(s.mint), model.Time(s.maxt))
 }