@codesome 2023-05-29
I modified `BenchmarkDistributorQuerier_Select` with this diff

<details>
  <summary>The diff</summary>

  ```diff
--- a/pkg/querier/distributor_queryable_test.go
+++ b/pkg/querier/distributor_queryable_test.go
@@ -8,6 +8,7 @@ package querier
 import (
 	"context"
 	"fmt"
+	"github.com/grafana/mimir/pkg/querier/batch"
 	"math"
 	"strconv"
 	"testing"
@@ -572,17 +573,27 @@ func BenchmarkDistributorQuerier_Select(b *testing.B) {
 	d.On("QueryStream", mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return(response, nil)
 
 	ctx := user.InjectOrgID(context.Background(), "0")
-	queryable := newDistributorQueryable(d, mergeChunks, newMockConfigProvider(0), nil, log.NewNopLogger())
+	queryable := newDistributorQueryable(d, batch.NewChunkMergeIterator, newMockConfigProvider(0), nil, log.NewNopLogger())
 	querier, err := queryable.Querier(ctx, math.MinInt64, math.MaxInt64)
 	require.NoError(b, err)
 
 	b.ResetTimer()
+	b.ReportAllocs()
 
 	for n := 0; n < b.N; n++ {
 		seriesSet := querier.Select(true, &storage.SelectHints{Start: math.MinInt64, End: math.MaxInt64})
 		if seriesSet.Err() != nil {
 			b.Fatal(seriesSet.Err())
 		}
+
+		var it chunkenc.Iterator
+		for seriesSet.Next() {
+			ss := seriesSet.At()
+			it = ss.Iterator(it)
+			if it.Err() != nil {
+				b.Fatal(it.Err())
+			}
+		}
 	}
 }
 
  ```
</details>

and here are the benchmark results. The big difference is likely because there is only 1 chunk with 1 sample. So it does not represent real-world load. But it confirms that the re-use is happening properly.

```
benchmark                                 old ns/op     new ns/op     delta
BenchmarkDistributorQuerier_Select-10     25454832      21200190      -16.71%

benchmark                                 old allocs     new allocs     delta
BenchmarkDistributorQuerier_Select-10     180129         90129          -49.96%

benchmark                                 old bytes     new bytes     delta
BenchmarkDistributorQuerier_Select-10     23246433      3300197       -85.80%
```