@charleskorn 2023-05-29
> Implements the re-use of iterator for `batch.NewChunkMergeIterator`. Although querier uses two other chunk iterators, since this one is enabled by default (and I guess most commonly), I gave this one a try first.

Under what circumstances would we want to enable the other kinds of iterators? (I'm wondering if we should separately deprecate and eventually remove the others)