@codesome 2023-05-30
> Under what circumstances would we want to enable the other kinds of iterators?

I personally don't know. I saw that this one is mostly used, so ignored the other two in the meantime.

> It's interesting that BenchmarkNewChunkMergeIterator_CreateAndIterate shows a slight CPU time degradation for the test cases with larger numbers of chunks - any ideas why this is the case?

The first run showed slight decrease, but it was slightly higher. Don't know the reason myself as well.