@pracucci 2023-05-30
> Under what circumstances would we want to enable the other kinds of iterators?

At Grafana Labs we only use the default `-querier.batch-iterators=true`. I'm also not aware of anyone in the community using any other iterator on purpose. The other iterators are there since years and I think we could just deprecated them and remove them (also I don't see much value spending time implementing the iterators reuse for them).